+++ web/static/js/datatable.js
@@ -1,4 +1,0 @@
-import $ from 'jquery'
-let dt = require( 'datatables.net-bs4' )( $ );
-
-$('.datatable').DataTable()