+++ web/static/js/socket.js
@@ -7,7 +7,7 @@ import {Socket} from "phoenix"
 
 let socket = new Socket("/socket", {params: {token: window.userToken }})
 
-
+// connect socket
 socket.connect()
 
 export default socket