+++ web/views/user_view.ex
@@ -1,11 +1,4 @@
 defmodule Collaboration.UserView do
   use Collaboration.Web, :view
 
-  def render("user-public.json", %{user: u}) do
-    %{
-      id: u.id,
-      firstname: u.firstname,
-      lastname: u.lastname
-    }
-  end
 end