+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 If versions are ommited from the changelog they only contain deployment updates
 
+### 0.2.20 - 08/24/2017
+  *  Improved: Much more pleasing Topic/Idea/Comment List
+
 ### 0.2.12 - 08/23/2017
   *  Fixed: Several Minor Bugs
   *  Fixed: Topics now only shows associated ideas (admin)